# api/openapi.yaml
openapi: 3.1.0
info:
  title: Hookyard API
  version: 0.1.0
  summary: Reliable delivery for outbound API calls.
  description: |
    Hookyard delivers outbound HTTP requests to third-party APIs ("upstreams") on behalf of your
    applications, with retries, backoff and a dead-letter queue.

    ## Concepts

    - **Upstream**: a third-party API configured in Hookyard (base URL, credentials, retry policy).
      Applications refer to upstreams by name and never handle vendor URLs or credentials.
    - **Request**: one outbound call that an application asked Hookyard to deliver.
    - **Attempt**: one HTTP try of a request.
    - **Dead-letter queue (DLQ)**: requests in the `dead` state that exhausted their retries or failed
      permanently. They can be inspected and replayed.

    ## Conventions

    - **Authentication**: every `/v1` endpoint requires `Authorization: Bearer <token>`.
    - **Durations** are strings with a unit, for example `"500ms"`, `"30s"`, `"5m"` or `"2h"`.
    - **Timestamps** are RFC 3339 in UTC.
    - **Errors** always use the `Error` schema, and `error.code` is stable and safe to branch on.
    - **Pagination** is cursor based: pass `next_cursor` from a response as `cursor` to get the next page.

    ## Delivery

    - Network errors, timeouts, `408`, `425`, `429` and `5xx` responses are retried with exponential
      backoff and full jitter. A `Retry-After` header on `429` or `503` sets the delay.
...
      moves to the dead-letter queue immediately.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/requests/{id}/callbacks:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    get:
      operationId: listCallbacks
      tags: [Callbacks]
      summary: List completion callbacks
      description: |
        Returns the callbacks sent (or being sent) for a request, oldest first. A request gets one each
        time it reaches a final state with a callback URL set; a replayed request can have several.
      responses:
        "200":
          description: The callbacks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallbackList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/InternalError"
  /v1/requests/{id}/callbacks/{callback_id}/retry:
    parameters:
      - $ref: "#/components/parameters/RequestId"
      - name: callback_id
        in: path
        required: true
        schema:
          type: string
          example: evt_0f6a3c2b9d8e4f5a8b7c6d5e4f3a2b1c
    post:
      operationId: retryCallback
      tags: [Callbacks]
      summary: Retry a failed callback
      description: |
        Sends a `failed` callback again, with a fresh attempt budget. The event (its id and body) is
        unchanged, so receivers that deduplicate on `webhook-id` handle it once. Recorded in the audit
        log.
      responses:
        "202":
          description: The callback was queued again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Callback"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API token configured in Hookyard.
  parameters:
...
    RequestId:
      name: id
      in: path
      required: true
      description: Request ID.
      schema:
        $ref: "#/components/schemas/RequestId"
...
  responses:
...
    Unauthorized:
      description: The bearer token is missing or invalid.
      headers:
        WWW-Authenticate:
          description: Always `Bearer realm="hookyard"`.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: unauthorized, message: missing or invalid bearer token }
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: not_found, message: "request \"req_01J9ZK3Q4W5E6R7T8Y9U0I1O2P\" not found" }
    Conflict:
      description: The resource is not in a state that allows this action.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: invalid_state, message: "cannot cancel a request that is in_flight" }
...
    InternalError:
      description: An unexpected server error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: internal, message: internal server error }
  schemas:
    RequestId:
      type: string
      description: Unique request ID, prefixed with `req_`.
      pattern: "^req_[0-9A-Za-z]+$"
      example: req_01J9ZK3Q4W5E6R7T8Y9U0I1O2P
...
    RequestStatus:
      type: string
      description: |
        - `scheduled`: waiting for `deliver_at`
        - `pending`: waiting to be picked up by a worker
        - `in_flight`: being delivered right now
        - `failed`: the last attempt failed and a retry is scheduled (see `next_attempt_at`)
        - `succeeded`: delivered successfully (final)
        - `dead`: retries exhausted or a permanent error; in the DLQ (final)
        - `unknown`: the request was sent but no response arrived, so the upstream may or may not have
          processed it; not retried to avoid a duplicate. Settle it with `resolve`, or `replay` it (final)
        - `canceled`: canceled before delivery (final)
      enum: [scheduled, pending, in_flight, failed, succeeded, dead, unknown, canceled]
...
    Callback:
      type: object
      description: One completion event and its delivery to the application.
      required: [id, request_id, type, url, request_status, status, attempt_count, created_at]
      properties:
        id:
          type: string
          description: The event id, sent as `webhook-id`. The same on every attempt.
          example: evt_0f6a3c2b9d8e4f5a8b7c6d5e4f3a2b1c
        request_id:
          $ref: "#/components/schemas/RequestId"
        type:
          $ref: "#/components/schemas/CallbackEventType"
        url:
          type: string
        request_status:
          $ref: "#/components/schemas/RequestStatus"
        status:
          type: string
          enum: [pending, delivering, delivered, failed]
          description: |
            `pending` (waiting for its first or next attempt), `delivering`, `delivered` (the app answered
            2xx) or `failed` (every attempt failed; retry it with `retryCallback`).
        attempt_count:
          type: integer
          minimum: 0
        next_attempt_at:
          type: [string, "null"]
          format: date-time
        last_status_code:
          type: [integer, "null"]
        last_error:
          type: [string, "null"]
        last_attempt_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        delivered_at:
          type: [string, "null"]
          format: date-time
    CallbackList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Callback"
    CallbackEventType:
      type: string
      enum: [request.succeeded, request.dead, request.unknown, request.canceled]
...
    ErrorCode:
      type: string
      description: Stable, machine-readable error code.
      enum:
        - bad_request
        - unauthorized
        - not_found
        - invalid_state
        - payload_too_large
        - validation_failed
        - unknown_upstream
        - internal
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: Human-readable explanation. Do not parse it; use `code`.
            details:
              type: array
              description: Per-field problems, for `validation_failed`.
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    description: Dotted path of the invalid field.
                    example: retry.max_attempts
                  message:
                    type: string